+++ contracts/test/TestRoot.t.sol
@@ -26,7 +26,7 @@ contract TestRoot is Test, Context {
         vm.startPrank(alice);
         vm.roll(1);
 
-        pool = new Pool("Pool", "P", int256(1e16));
+        pool = new Pool("Pool", "P", 1e16);
         tokens = new Token[](NTOKENS);
         addresses = new address[](NTOKENS);
