+++ contracts/test/Users.t.sol
@@ -18,7 +18,7 @@ contract UsersTest is Test {
         vm.startPrank(alice);
         vm.roll(1);
 
-        pool = new Pool("Pool", "P", int256(1e16));
+        pool = new Pool("Pool", "P", 1e16);
     }
 
     function testUsers_addUser() public {
